refactor: callbacks
